requirements.txt
Django==4.0.2
djangorestframework==3.13.1
django-environ==0.8.1
Pillow==9.0.1
coreapi==2.3.3
redis==4.1.4
hiredis==2.0.0
django-filter==21.1
django-phonenumber-field==6.1.0
